encrypt password and account number * the encryption uses a unique machine key, this means that if anyone steals config.otml with a saved password, he will not be able to decrypt the password without the machine UUID key * the encrypt uses a simple XOR encryption method, encoded with base64 and adler32 summing

+++ src/framework/stdext/string.h
@@ -161,7 +161,7 @@ inline std::string date_time_string() {
 }
 
 /// Convert decimal to hexadecimal
-inline std::string dec_to_hex(unsigned int num) {
+inline std::string dec_to_hex(uint64 num) {
     std::string str;
     std::ostringstream o;
     o << std::hex << num;
@@ -170,8 +170,8 @@ inline std::string dec_to_hex(unsigned int num) {
 }
 
 /// Convert hexadecimal to decimal
-inline unsigned int hex_to_dec(const std::string& str) {
-    unsigned int num;
+inline uint64 hex_to_dec(const std::string& str) {
+    uint64 num;
     std::istringstream i(str);
     i >> std::hex >> num;
     return num;
